Moved CPAR elements to prep for CPAR display
